ml-api/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
scikit-learn
pandas
python-multipart
google-genai
python-dotenv
shap
torch
dowhy
matplotlib